backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
faster-whisper==1.0.0
demucs==4.0.1
edge-tts==6.1.9
ffmpeg-python==0.2.0
openai==1.10.0
pydantic==2.6.0
loguru==0.7.2
python-dotenv==1.0.1
websockets==12.0
requests==2.31.0
